Consolidate branch merging into `Branch` objects to maintain consistent `unmerged` status information (#1028)
* Generalize listening for all (instead of selected) rejected actions in listenerMiddleware

* Fixes for regex patterns to handle branch comparisons and entry formats in worktreeStatus command

* Throw console.error if file-only FSCache observes an unlinkDir event

* FSCache updates to track card UUIDs in 'reserved' instead of metafile UUIDs

* Update status of all modified sub-files when calling worktreeStatus in updateVersionedMetafile command

* Remove unnecessary imports

* Rebuild yarn.lock

* Resolve absolute paths based on repository root when filtering worktreeStatus entries in fileStatus command

* Observe updated metafile for each update step when listening to cardAdded or cardUpdated actions

* Translate 'uncommitted' branch status into '*modified' on directories for fileStatus command

* Combine 'revertStagedChanges' and 'revertUnmergedChanges' into a single 'revertChanges' async thunk for metafiles

* Fix newline parsing error in JSDoc block comments

* Convert namespace import to named imports for slices/metafiles

* Remove useEffect for hydrating child metafiles in Directory since listenerMiddleware already handles filebased and versioned updates

* Reset all unstaged, staged, and unmerged files using 'revertChanges' thunk in ResetButton

* Expand StyledTreeComponent to include a onHoverText for adding version status on hover of FileComponent in Explorer cards

* Filter by git status instead of conflicts array in selectByConflicted selector

* ConflictManagerReverse component for displaying UUID, Path, Update, Repo, Status, and Branch fields

* Updated listenerMiddleware for card additions/updates/removals that impact FSCache subscriptions

* Log action.meta fields in rejected async thunk actions observed by listenerMiddleware

* Populate branch field to align ConflictManager metafile types at creation and response callsites

* Filter conflicting metafiles by handler type instead of using isFilebasedMetafile since conflict metafiles might not be populated with a content field

* Metafile 'loading' field renamed to 'flags' and extracted into Flag type

* Transition to using the nullish coalescing operator (??) from TS 3.7

* GitStatus filters for staged/unstaged/conflicted consolidated into utils

* Branch motif colors expanded to indicate branches with merge conflicts

* Check branch status in updateBranch thunk

* Badge component to display number of conflicting chunks in Card components containing textual content

* Remove redundant check for unmerged conflict metafiles in updateVersionedMetafile since worktreeStatus is capable of detecting unmerged status

* Update metafile with new code content using nullish coalescing operator (??) from TS 3.7

* Remove redundant updateBranch call during merge since updateBranches now calls updateBranch internally

* Rename utils.isConflicted to isUnmerged to indicate that unmerged status differs from containing conflicting chunks

* Await updateBranches immediately upon Refresh button click event

* Extract regex patterns for conflicting chunks into utils.getConflictingChunks

* Check for conflicting chunks when determining if conflicted motif should be returned in getSourceMotif since status can change while chunks remain in content

* JSDoc @returns type for flattenArray and flattenObject

* symmetrical fn for partitioning two arrays into symmetrical differences and intersection subarrays

* updateVersionedMetafile thunk updates descendants for directory parameter metafiles

* selectByConflicted selector expanded to differentiate metafiles based on branches

* JSDoc @returns type for diff

* JSDoc @returns type for useMap

* JSDoc @returns type for usePrevious

* Newline after description in JSDoc block

* JSDoc @returns types and newlines after description blocks for immutables fns

* Newline after description in JSDoc blocks and adhering to jsdoc/multiline-blocks rule

* JSDoc @param and @returns types for createState and mockStore

* JSDoc comments clarified to indicate listeners that match on async thunk actions

* Fix listener updates for fulfilled Metafile async thunk update actions to use updated versions of underlying metafile

* Rename isUnstaged() to isModified() to more accurately reflect a modified, but not staged status

* Update conflicts field during onChange event in Editor

* Remove unused import withStyles

* All Button components updated to prevent propogating click events to parent components

* Refresh button component for updating filebased and versioned metafile fields

* BranchRibbon component responds to conflicts by changing text and background colors

* Add `electron-devtools-installer` package to devDependencies to auto install DevTool extensions into Electron

* Enable `redux-devtools` module in Electron app

* Remove unused module.exports.fallback imports

* Remove `plugin:import/electron` and use`settings.import/modules: ["electron"]` instead

* Standardized quotes in jest.config.js moduleNameMapper

* Updated ARCHITECTURE for changes to import packages and TS/ESLint/Jest/RTL configs

* ConflictManager card component removed, conflicts managed directly in Explorer card component

* Add `enabled` prop for enabling/disabling Button components during render

* Remove listener checks for deprecated `ConflictManager` metafile types

* Allow multiple astericks in JSDoc comments for Markdown bullet lists

* Fix overflowing card-header when title is long

* New Metadata component for a standardized table view of Metafile contents on the back of cards

* BranchList component simplified and migrated into Branches component tree

* BranchRibbon component simplified and migrated into Branches component tree

* JSDoc comments for Branches component

* BranchItem component allows highlights for indicating the current branch of a parent card, metafile, or list

* BranchesView button component for selecting Branches view on the back of cards

* Migrating Card components to use Metadata component on reverse

* Editor code content switched to pass-through from Metafile content property

* Updated ref to BranchRibbon component

* Updated jsdoc/no-multi-astericks rule format

* removeObjectProperty function added for immutably removing properties in a typesafe manner

* worktreeStatus checks for in-progress merges and returns 'unmerged' if found

* Branch objects expanded to include merging field when a branch is the base of an in-progress merge

* Metafile object no longer differentiates merging status, only conflict status (Branch objects handle merging now)

* BranchRibbon indicates unmerged based on branch.status instead of conflicted metafile count

* Abort button enable via checks against merging branches instead of merging metafiles

* Resolve button simplified to key off of merging branches in Explorer metafiles

* Stage button checks for unstaged metafiles based on unmerged status instead of modified status

* Remove unused isAbortable import

* Metadata component checks for merging branches instead of merging metafiles

* MergeDialog modal component checks for merging branches and updates accordingly

* fetchMergingBranches checks for the existence of MERGE_MSG file before attempting to read

* Remove unused SHA1 type import

* Always enable Redux DevTools integration

* Mock git-status for all fetchBranch tests that intrinsically check branch status

* Mock git-status for all updateVersionedMetafile tests that intrinsically check branch status

* Partial removal of strict RootState typing on useAppSelector usage

+++ src/store/listenerMiddleware.ts
@@ -1,90 +1,95 @@
-import { createListenerMiddleware, addListener, TypedStartListening, TypedAddListener, isRejected, isAnyOf, isPending } from '@reduxjs/toolkit';
+import { addListener, createListenerMiddleware, isAnyOf, isFulfilled, isRejected, TypedAddListener, TypedStartListening } from '@reduxjs/toolkit';
 import { DateTime } from 'luxon';
+import branchSelectors from './selectors/branches';
 import cacheSelectors from './selectors/cache';
 import cardSelectors from './selectors/cards';
 import metafileSelectors from './selectors/metafiles';
 import { cacheRemoved } from './slices/cache';
 import { cardAdded, cardRemoved, cardUpdated } from './slices/cards';
-import { isDirectoryMetafile, isFilebasedMetafile, isFileMetafile, metafileUpdated } from './slices/metafiles';
-import { RootState, AppDispatch } from './store';
-
-import { subscribe, unsubscribe } from './thunks/cache';
-import { createMetafile, fetchMetafile, fetchParentMetafile, updateVersionedMetafile, updateFilebasedMetafile, switchBranch } from './thunks/metafiles';
-import { fetchRepo, buildRepo } from './thunks/repos';
-import { UUID } from './types';
-import branchSelectors from './selectors/branches';
+import { isDirectoryMetafile, isFilebasedMetafile, isFileMetafile, metafileRemoved, metafileUpdated } from './slices/metafiles';
+import { AppDispatch, RootState } from './store';
 import { removeBranch } from './thunks/branches';
+import { subscribe, unsubscribe } from './thunks/cache';
+import { switchBranch, updateFilebasedMetafile, updateVersionedMetafile } from './thunks/metafiles';
+import { UUID } from './types';
 
 export const listenerMiddleware = createListenerMiddleware<RootState>();
-
 export type AppStartListening = TypedStartListening<RootState, AppDispatch>;
-
 export const startAppListening = listenerMiddleware.startListening as AppStartListening;
-
 export const addAppListener = addListener as TypedAddListener<RootState, AppDispatch>;
-
-const isRejectedAction = isRejected(fetchMetafile, createMetafile, updateFilebasedMetafile, updateVersionedMetafile, fetchParentMetafile, fetchRepo, buildRepo);
 
 /**
  * Listen for rejected actions and log relevant information in the console.
  */
 startAppListening({
-    matcher: isRejectedAction,
+    matcher: isRejected,
     effect: async (action) => {
         console.group(`${action.type} : ${DateTime.local().toHTTP()}`);
+        console.log(`meta: `, action.meta);
         console.log(action.error);
         console.groupEnd();
     }
 });
 
 /**
- * Listen for pending updates to Metafiles to toggle `loading` flags for UI indicators.
+ * Listen for pending Metafile async thunk update actions and add the `updating` flag.
  */
 startAppListening({
-    matcher: isAnyOf(updateVersionedMetafile.pending, updateVersionedMetafile.fulfilled, updateVersionedMetafile.rejected),
+    matcher: isAnyOf(updateFilebasedMetafile.pending, updateVersionedMetafile.pending),
     effect: async (action, listenerApi) => {
-        if (isPending(updateVersionedMetafile)(action)) {
-            listenerApi.dispatch(metafileUpdated({ ...action.meta.arg, loading: [...action.meta.arg.loading, 'versioned'] }));
-        }
-        if (isAnyOf(updateVersionedMetafile.fulfilled, updateVersionedMetafile.rejected)(action)) {
-            const metafile = listenerApi.getState().metafiles.entities[action.meta.arg.id];
-            if (metafile) listenerApi.dispatch(metafileUpdated({ ...metafile, loading: action.meta.arg.loading.filter(flag => flag !== 'versioned') }));
+        if (isAnyOf(updateFilebasedMetafile.pending, updateVersionedMetafile.pending)(action) && action.meta.arg.flags.every(flag => flag !== 'updating')) {
+            listenerApi.dispatch(metafileUpdated({ ...action.meta.arg, flags: [...action.meta.arg.flags, 'updating'] }));
         }
     }
 });
+
+/**
+ * Listen for fulfilled/rejected Metafile async thunk update actions and remove the `updating` flag.
+ */
+startAppListening({
+    matcher: isAnyOf(updateFilebasedMetafile.fulfilled, updateVersionedMetafile.fulfilled, updateFilebasedMetafile.rejected, updateVersionedMetafile.rejected),
+    effect: async (action, listenerApi) => {
+        if (isFulfilled(updateFilebasedMetafile, updateVersionedMetafile)(action)) {
+            listenerApi.dispatch(metafileUpdated({ ...action.payload, flags: action.meta.arg.flags.filter(flag => flag !== 'updating') }));
+        }
+        if (isRejected(updateFilebasedMetafile, updateVersionedMetafile)(action)) {
+            listenerApi.dispatch(metafileUpdated({ ...action.meta.arg, flags: action.meta.arg.flags.filter(flag => flag !== 'updating') }));
+        }
+    }
+});
+
+/**
+ * Listen for pending Branch switches and add the `checkout` flag to all related Metafiles.
+ */
+startAppListening({
+    actionCreator: switchBranch.pending,
+    effect: async (action, listenerApi) => {
+        const branch = branchSelectors.selectByRef(listenerApi.getState(), action.meta.arg.ref)[0];
+        const metafiles = branch ? metafileSelectors.selectByBranch(listenerApi.getState(), branch.id, action.meta.arg.root) : [];
+        metafiles
+            .filter(metafile => metafile.flags.every(flag => flag !== 'checkout'))
+            .map(metafile => listenerApi.dispatch(metafileUpdated({ ...metafile, flags: [...metafile.flags, 'checkout'] })));
+    }
+});
+
 
 /**
  * Listen for switch branch actions and update Metafiles to toggle `checkout` flags for UI indicators.
  */
 startAppListening({
-    matcher: isAnyOf(switchBranch.pending, switchBranch.fulfilled, switchBranch.rejected),
+    matcher: isAnyOf(switchBranch.fulfilled, switchBranch.rejected),
     effect: async (action, listenerApi) => {
-        if (isPending(switchBranch)(action)) {
-            const branch = branchSelectors.selectByRef(listenerApi.getState(), action.meta.arg.ref)[0];
-            const metafiles = branch ? metafileSelectors.selectByBranch(listenerApi.getState(), branch.id, action.meta.arg.root) : [];
-            metafiles.map(metafile => listenerApi.dispatch(metafileUpdated({ ...metafile, loading: [...metafile.loading, 'checkout'] })));
-        }
         if (isAnyOf(switchBranch.fulfilled, switchBranch.rejected)(action)) {
             const branch = branchSelectors.selectByRef(listenerApi.getState(), action.meta.arg.ref)[0];
             const metafiles = branch ? metafileSelectors.selectByBranch(listenerApi.getState(), branch.id, action.meta.arg.root) : [];
-            metafiles.map(metafile => listenerApi.dispatch(metafileUpdated({ ...metafile, loading: metafile.loading.filter(flag => flag !== 'checkout') })));
+            metafiles
+                .map(metafile => listenerApi.dispatch(metafileUpdated({ ...metafile, flags: metafile.flags.filter(flag => flag !== 'checkout') })));
         }
     }
 });
 
 /**
- * Listen for card removal actions and remove any Diff cards that reference it.
- */
-startAppListening({
-    actionCreator: cardRemoved,
-    effect: async (action, listenerApi) => {
-        const diffs = cardSelectors.selectByTarget(listenerApi.getState(), action.payload as UUID);
-        diffs.map(card => listenerApi.dispatch(cardRemoved(card.id)));
-    }
-});
-
-/**
- * Listen for branch removal actions and remove any cards that reference it.
+ * Listen for branch removal actions and remove any cards that reference that branch.
  */
 startAppListening({
     actionCreator: removeBranch.fulfilled,
@@ -97,7 +102,7 @@ startAppListening({
 });
 
 /**
- * Listen for added or switched metafiles for Cards to update versioned fields, directory fields, and subscriptions.
+ * Listen for card adding/switching actions and update child Metafiles in directories, subscriptions, and cache adding.
  */
 startAppListening({
     predicate: (action, _, previousState) => {
@@ -109,49 +114,51 @@ startAppListening({
         return false;
     },
     effect: async (action, listenerApi) => {
-        const metafile = metafileSelectors.selectById(listenerApi.getState(), action.payload.metafile);
+        if (cardAdded.match(action) || cardUpdated.match(action)) {
+            const metafile = listenerApi.getState().metafiles.entities[action.payload.metafile];
 
-        if (metafile && (cardAdded.match(action) || cardUpdated.match(action))) {
+            if (metafile && isFilebasedMetafile(metafile)) {
+                let updated = await listenerApi.dispatch(updateFilebasedMetafile(metafile)).unwrap();
+                updated = await listenerApi.dispatch(updateVersionedMetafile(updated)).unwrap();
+                const metafiles = isDirectoryMetafile(updated) ? metafileSelectors.selectByIds(listenerApi.getState(), updated.contains) :
+                    isFileMetafile(updated) ? metafileSelectors.selectByIds(listenerApi.getState(), [metafile.id]) : [];
 
-            if (isFilebasedMetafile(metafile)) {
-                const updated = await listenerApi.dispatch(updateFilebasedMetafile(metafile)).unwrap();
-                await listenerApi.dispatch(updateVersionedMetafile(metafile));
-
-                if (isDirectoryMetafile(updated)) {
-                    const children = metafileSelectors.selectByIds(listenerApi.getState(), updated.contains);
-                    await Promise.all(children.filter(isFilebasedMetafile).map(child =>
-                        listenerApi.dispatch(updateFilebasedMetafile(child))
-                    ));
-                }
-
-                if (isFileMetafile(updated)) {
-                    await listenerApi.dispatch(subscribe({ path: updated.path.toString(), metafile: updated.id }));
-                }
+                metafiles.filter(isFileMetafile).forEach(async metafile => {
+                    await listenerApi.dispatch(subscribe({ path: metafile.path.toString(), card: action.payload.id }));
+                });
             }
         }
-    },
+    }
 });
 
 /**
- * Listen for card removal actions and update child Metafiles in directories, subscriptions, and cache removals
+ * Listen for card removal actions and update child Metafiles in directories, subscriptions, and cache removals.
+ * Also remove any Diff cards that reference that card.
  */
 startAppListening({
     actionCreator: cardRemoved,
     effect: async (action, listenerApi) => {
-        const state = listenerApi.getState();
         const card = listenerApi.getOriginalState().cards.entities[action.payload];
-        const metafile = card ? state.metafiles.entities[card.metafile] : undefined;
-        if (card && metafile) {
-            const childIds = isDirectoryMetafile(metafile) ? metafile.contains :
-                isFileMetafile(metafile) ? [metafile.id] : [];
-            const metafiles = metafileSelectors.selectByIds(state, childIds);
-            metafiles.filter(isFilebasedMetafile).forEach(metafile => {
-                const existing = cacheSelectors.selectById(state, metafile.path.toString());
+        const metafile = card ? listenerApi.getState().metafiles.entities[card.metafile] : undefined;
+
+        // handle Diff cards that contain a reference to the removed card
+        const diffs = cardSelectors.selectByTarget(listenerApi.getState(), action.payload as UUID);
+        diffs.map(card => listenerApi.dispatch(cardRemoved(card.id)));
+
+        if (card && metafile && isFilebasedMetafile(metafile)) {
+            let updated = await listenerApi.dispatch(updateFilebasedMetafile(metafile)).unwrap();
+            updated = await listenerApi.dispatch(updateVersionedMetafile(updated)).unwrap();
+            const metafiles = isDirectoryMetafile(updated) ? metafileSelectors.selectByIds(listenerApi.getState(), updated.contains) :
+                isFileMetafile(updated) ? metafileSelectors.selectByIds(listenerApi.getState(), [metafile.id]) : [];
+
+            metafiles.filter(isFileMetafile).forEach(metafile => {
+                const existing = cacheSelectors.selectById(listenerApi.getState(), metafile.path.toString());
                 if (existing) {
                     if (existing && existing.reserved.length > 1) {
-                        listenerApi.dispatch(unsubscribe({ path: existing.path, metafile: metafile.id }));
+                        listenerApi.dispatch(unsubscribe({ path: existing.path, card: card.id }));
                     } else {
                         listenerApi.dispatch(cacheRemoved(existing.path));
+                        listenerApi.dispatch(metafileRemoved(metafile.id));
                     }
                 }
             });
